Add define statement for sizeX and sizeY
A define statement was added for both sizeX and sizeY to define the size
that will be used for the board.

Previously sizeX and sizeY was defined using a const int value but this
caused a compilation error when defining arrays with the values of sizeX
and sizeY. Changing the values to be defined globally with a #define
fixed the issue and lets the game function normally.